% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_description_utils.R
\name{proj_desc}
\alias{proj_desc}
\title{usethis project description functions}
\usage{
proj_desc(path = proj_get())
}
\arguments{
\item{path}{file path}
}
\description{
Quiet functions for our project
}
