% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_add_author_test.R
\name{psa_add_author}
\alias{psa_add_author}
\title{Create PSA Project}
\usage{
psa_add_author(
  given = NULL,
  family = NULL,
  ORCID_number = NULL,
  email = NULL,
  role = NULL
)
}
\arguments{
\item{path}{path to directory where the local project will be}

\item{psa_name}{name of the repo (e.g., PSA000-demo)}
}
\description{
DESCRIPTION
ADD INFORMATION ABOUT \\ OR / FOR WINDOWS
}
\keyword{create}
\keyword{project}
