% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_add_documents.R
\name{psa_add_documents}
\alias{psa_add_documents}
\title{Add Documents to Project}
\usage{
psa_add_documents(
  folder = NULL,
  path = NULL,
  should_replace = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{folder}{the folder to be added.
\itemize{
\item "ethics": creates the "01_Ethics" folder.
\item "power": creates the "02_Power" folder.
\item "materials": creates the "03_Materials" folder.
\item "procedure": creates the "04_Procedure" folder.
\item "data": creates the "05_Data" folder.
\item "analysis": creates the "06_Analysis" folder.
\item "communication": creates the "07_Communication
folder. Any other parameter option will be added
into the 08_Other folder.
}}

\item{path}{path of the requested file/directory.
Must be a valid path within the current device.}

\item{should_replace}{whether the requested file
must overwrite an already existing one within the
project.}

\item{recursive}{whether directories should be
copied recursively.}
}
\description{
This function copies the requested file/directory
to the specified folder within the project folder
structure. If the requested folder does not already
exist, it is created automatically.
}
\examples{
\dontrun{
psa_add_documents(folder = "ethics",
  path = "~/Downloads")
psa_add_documents(folder = "ethics",
  path = "~/Downloads",
  recursive = TRUE)
}
}
\keyword{add}
\keyword{documents}
