% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_check_folder.R
\name{psa_check_folder}
\alias{psa_check_folder}
\title{Create PSA Project}
\usage{
psa_check_folder(folder = "all")
}
\arguments{
\item{path}{path to directory where the local project will be}

\item{psa_name}{name of the repo (e.g., PSA000-demo)}
}
\description{
DESCRIPTION
ADD INFORMATION ABOUT \\ OR / FOR WINDOWS
}
\keyword{create}
\keyword{project}
