% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_citation.R
\name{psa_citation}
\alias{psa_citation}
\title{Create PSA Citation for Project}
\usage{
psa_citation(
  bibtype = "Article",
  title = "desc",
  author = "desc",
  journal = NULL,
  year = NULL,
  url = NULL,
  ...
)
}
\arguments{
\item{bibtype}{a character string with a BibTeX entry type.
See Entry Types for details.}

\item{title}{the title of the manuscript and/or title someone
should use to cite this resource.}

\item{author}{an author entry to include the author for citation.
Please see the build project and/or update description for
how to format author entries.}

\item{journal}{The journal or other location to cite for the
project.}

\item{year}{The year of the citation for the project.}

\item{url}{The url to include for the citation of the project.}

\item{...}{Other information that can be passed to bibentry to
add more details for the citation of the project.}
}
\description{
Allows the user to create the citation for the github and/or
overall project for users to cite the data and files.
}
\keyword{citation}
\keyword{create}
