% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_create_description.R
\name{psa_create_description}
\alias{psa_create_description}
\title{Create PSA Project Description}
\usage{
psa_create_description(
  package_title = NULL,
  description = NULL,
  version = NULL,
  license = NULL
)
}
\arguments{
\item{path}{path to directory where the local project will be}

\item{psa_name}{name of the repo (e.g., PSA000-demo)}
}
\description{
Create PSA Project Description
}
\examples{
# psa_create_description(package_title = "PSA000-demo",
#   description = "Here's our description",
#   version = "0.0.1")
}
\keyword{create}
\keyword{description}
