% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_dependencies.R
\name{psa_dependencies}
\alias{psa_dependencies}
\title{Add R package dependencies for analysis code}
\usage{
psa_dependencies(scan_folder = "inst", add_package = NULL)
}
\arguments{
\item{scan_folder}{which folder should be examined
for .R or .Rmd files to find package libraries}

\item{add_package}{Add a single or set of packages
to the project, use c("package", "package") to
include several at once.}
}
\description{
This function will look at all .R or .Rmd files for
R packages to include in the package dependencies list.
You can tell it to scan a specific folder - or use the
default \code{inst} - or add a package directly. If you
use the \code{add_package} option, the \code{scan_folder} will
be ignored.
}
\examples{
# psa_dependencies(scan_folder = "inst")
}
\keyword{dependencies}
