% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_git.R
\name{psa_git}
\alias{psa_git}
\title{Add R package to PsySciAcc Organization}
\usage{
psa_git(organisation = "psysciacc", private = TRUE)
}
\arguments{
\item{organisation}{should be "psysciacc" - please
note you should confirm you have the privileges
to push to the organization first}

\item{private}{a TRUE/FALSE variable with the
default to start as private until you are ready
to release the project package.}
}
\description{
This function will initialize the use of github
and direct the package to be included in the
PsySciAcc organization. You can then use RStudio's
built in git push/pull, the command line, or your
favorite gui desktop app to push and pull to
github.
}
\examples{
# psa_git()
}
\keyword{github}
