% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_update_description.R
\name{psa_update_description}
\alias{psa_update_description}
\title{Create/Update PSA Project Description}
\usage{
psa_update_description(
  package_title = NULL,
  description = NULL,
  version = NULL,
  license = NULL
)
}
\arguments{
\item{package_title}{new title for the package}

\item{description}{short description of package functionality}

\item{version}{version number of package (by default 0.0.0.9000)}

\item{license}{license type of package. Directly available options:
\itemize{
\item MIT License: "MIT"
\item Apache License 2.0: "Apache"
\item GNU General Public License: "GPLv2" or "GPLv3"
\item GNU Affero General Public License, version 3: "AGPLv3"
\item GNU Lesser General Public License: "LGPLv2.1" or "LGPLv3"
\item Creative Commons CC0 Public Domain Dedication: "CC0"
\item Creative Commons Attribution 4.0 International: "CC-BY"
\item Proprietary license: "Proprietary"
}

Refer to the documentation on package licenses in the usethis package website for available licenses.}
}
\description{
Create/Update PSA Project Description
}
\examples{
\dontrun{
psa_update_description(
 package_title = "New Title",
 description = "Short description",
 version = "0.0.1.0000",
 license = "MIT"
)
}
#
}
\keyword{create}
\keyword{description}
\keyword{update}
