% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_view_folder.R
\name{psa_view_folder}
\alias{psa_view_folder}
\title{View metadata for a folder}
\usage{
psa_view_folder(folder = NULL)
}
\arguments{
\item{folder}{what folder you want to view metadata for.
Any of the following names can be used:
\itemize{
\item "ethics": "01_Ethics",
\item "power": "02_Power"
\item "materials": "03_Materials"
\item "procedure": "04_Procedure"
\item "data": "05_Data"
\item "analysis": "06_Analysis"
\item "communication": "07_Communication"
\item "other": "08_Other"
}}
}
\value{
a dataframe of file information for the folder
}
\description{
Allows you to view the metadata for a folder for
a PSA project.
}
\examples{
\dontrun{
psa_view_folder(
folder = "ethics"
}


}
\keyword{metadata}
\keyword{view}
