% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psaverse-package.R
\docType{package}
\name{psaverse-package}
\alias{psaverse}
\alias{psaverse-package}
\title{psaverse: Archive PSA Projects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package will facilitate archiving and sharing all of the data and materials for Psychological Science Accelerator projects in a standardized, transparent way.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://psysciacc.org/psaverse/}
}

}
\author{
\strong{Maintainer}: Lisa DeBruine \email{debruine@gmail.com} (\href{https://orcid.org/0000-0002-7523-5539}{ORCID})

Authors:
\itemize{
  \item Marton Kovacs \email{marton.balazs.kovacs@gmail.com} (\href{https://orcid.org/0000-0002-8142-8492}{ORCID})
  \item Erin Buchanan \email{buchananlab@gmail.com} (\href{https://orcid.org/0000-0002-9689-4189}{ORCID})
  \item Ekatarina Pronizius \email{katjapron@gmail.com} (\href{https://orcid.org/0000-0003-1446-196X}{ORCID})
}

}
\keyword{internal}
