% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_add_author.R
\name{psa_add_author}
\alias{psa_add_author}
\title{Add Author Info to PSA Project}
\usage{
psa_add_author(
  given = NULL,
  family = NULL,
  ORCID_number = NULL,
  email = NULL,
  role = NULL
)
}
\arguments{
\item{given}{given name of author}

\item{family}{family name of author}

\item{ORCID_number}{ORCID number associated with author}

\item{email}{email address of author}

\item{role}{vector containing roles of author in the project, usually "aut" and "cre"; refer to utils::person() documentation for more information.}
}
\description{
Adds a person information to the \code{Authors@R} section of the DESCRIPTION
file, creating that field if necessary. Acts as a wrapper for the function
usethis::use_author() function.
}
\examples{
\dontrun{
psa_add_author(
  given = "Lucy",
  family = "van Pelt",
  role = c("aut", "cre"),
  email = "lucy@example.com",
  comment = c(ORCID = "LUCY-ORCID-ID")
}


}
\keyword{add}
\keyword{author}
