% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_check_folder.R
\name{psa_check_folder}
\alias{psa_check_folder}
\title{Check PSA Project Folder Structure}
\usage{
psa_check_folder(folder = "all")
}
\arguments{
\item{folder}{name of the folder to check (or "all") to check all folders.}
}
\description{
Checks the current project's folder structure according to the PSAVerse
standard structure, either one project folder (i.e., 01_Ethics, 02_Power,
etc.) at a time, or everything at once.
}
\keyword{check}
\keyword{folder}
