% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_create_metadata.R
\name{psa_create_metadata}
\alias{psa_create_metadata}
\title{Create Metadata for each folder}
\usage{
psa_create_metadata(folder, overwrite = FALSE)
}
\arguments{
\item{folder}{name of the folder to be documented. If not "ethics", "power",
..., "other", must supply full path of the folder.}

\item{overwrite}{TRUE/FALSE variable for if you want to overwrite a
previously created metadata file}
}
\description{
Input: a folder in inst.
Output: a README with info about the files in the folder, and templates for
the author to insert description.
}
\keyword{create}
\keyword{folder}
\keyword{metadata}
