% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_create_project.R
\name{psa_create_project}
\alias{psa_create_project}
\title{Create PSA Project}
\usage{
psa_create_project(
  path = "./",
  psa_name = "PSA000-demo",
  organisation = "psysciacc"
)
}
\arguments{
\item{path}{path to directory where the local project will be}

\item{psa_name}{name of the repo (e.g., PSA000-demo)}

\item{organisation}{github org to create the repo}
}
\description{
Create PSA Project
}
