% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_create_project.R
\name{psa_create_project}
\alias{psa_create_project}
\title{Create PSA Project}
\usage{
psa_create_project(
  path = "./",
  psa_name = "PSA000-demo",
  given = NULL,
  family = NULL,
  email = NULL,
  role = NULL,
  comment = NULL,
  open = FALSE
)
}
\arguments{
\item{path}{path to directory where the local project will be}

\item{psa_name}{name of the repo (e.g., PSA000-demo)}

\item{given}{given name of first author}

\item{family}{family name of first author}

\item{email}{email address of first author}

\item{role}{role of first author, usually "aut" and "cre";
refer to the documentation for utils::person() for more details.
You may only have one "cre" (maintainer).}

\item{comment}{additional information, e.g. ORCID number}

\item{open}{Defaults to FALSE, determine if you want to
immediately open the new package in an RStudio window.}
}
\description{
Use this function to initialize a PSA structured R project
to put on GitHub and share. Remember that file paths need
to be either "C:/Windows/blah" or "C:\\Windows\\blah" if you
are using a Windows machine.
}
\examples{
\dontrun{
psa_create_project(path = "~/Downloads", psa_name = "PSA001-Faces")
psa_create_project(path = "C:\\\\Downloads", psa_name = "PSA001-Faces")
}
}
\keyword{create}
\keyword{project}
